ENH: Changing the MAX temperature threshold to 90 c
The reason this change was made is because during testing we saw that
at 70c ambient the temperature at the sensor was very close to 80c. This
test was done on 10 units in a thermal chamber.

NOTE: This change was requested by Aviram even though the components are
rated at 85c. Since we are NOT doing a moving average of the temperatures
it may be justified.