BUG: DO not do a watchdog reset if in debug mode i.e usb ID GPIO is low
In MCU version a.1.6.0 we implemented a monitor to verify whether the OS booted up correctly. If the device had OS21 and below, the monitor would fail and the device would keep resetting. The problem was that there was no way to update the OS.

To fix this, if the device is in debug mode where USB ID GPIO is low, we
disable watchdog resets. This allows the user to update the OS even though
the MCU has not got a correct boot up notification (via the CPU_STATUS
GPIO)