ENH: Disable interrupt based OTG_ID monitoring and use polling
With new inthinc OTG_ID cables we were seeing a lot of noise on the
OTG_ID signal when UART data was transmitted. To avoid false low states
we switched from an interrupt edge trigerred approach to a polling
based approach. We store the previous state, so we don't keep configuring
GPIOs that have already been configured.